Change Nuget path to use SolutionDir
When this repo is added as a submodule to another project the providers don't build correctly because the project looks for the NuGet files in a packages folder by relative path. Since the packages folder is created in the solution dir it would be better to look for the files using the $(SolutionDir) variable.